add auth handler support (just twitch for now)
Access via the `--auth BACKEND` flag, which will trigger the auth
flow.

Currently only exists for twitch, but the hooks are there if an
auth flow is needed for some other backend.

Need to work on the extra information handling at some point, though.
Right now the main code accepts a -t flga for the twitch token, but
really any backend could (and probably would) have an oauth token or
something similar that would need to be stored, so we should handle
that.